Fix issues with capturing certain commands
This uses a sed-based regex instead of the
read builtin to drop the initial number
output by the 'history' command.

The problems with the read builtin are:
  - It only reads a single line.
  - It strips whitespace.

Changing to this sed expression (which
removes the initial number from the first
line only, assuming that the format of
the history line is "%5d%c %s%s\n" as
in history.def in the bash source code)
resolves both of these issues,
fixing #77 and #78 .

This also removes the need for either an
echo or a printf, and so also incidentally
resolves #76 . (As pointed out on that
issue, this could also have been fixed
without switching to sed.)

Adds regression tests for the fixed issues.
The test code is also adjusted to not have the
same echo/printf issue.